starting user dashboard

+++ functions/APIs/todos.js
@@ -4,6 +4,7 @@ const { db } = require("../util/admin");
 
 exports.getAllTodos = (request, response) => {
   db.collection("todos")
+    .where("username", "==", request.user.username)
     .orderBy("createdAt", "desc")
     .get()
     .then((data) => {
@@ -24,6 +25,21 @@ exports.getAllTodos = (request, response) => {
     });
 };
 
+exports.getOneTodo = (request, response) => {
+  let document = db.doc(`/todos/${request.params.todoId}`);
+  document
+    .get()
+    .then((doc) => {
+      if (!doc.exists) {
+        return response.status(404).json({ error: "Todo not found" });
+      }
+      return response.status(200).json(doc.data());
+    })
+    .catch((err) => {
+      console.log("Error getting document", err);
+    });
+};
+
 exports.postOneTodo = (request, response) => {
   if (request.body.body.trim() === "") {
     return response.status(400).json({ body: "Must not be empty" });
@@ -36,6 +52,7 @@ exports.postOneTodo = (request, response) => {
   const newTodoItem = {
     title: request.body.title,
     body: request.body.body,
+    username: request.user.username,
     createdAt: new Date().toISOString(),
   };
   db.collection("todos")
@@ -51,21 +68,49 @@ exports.postOneTodo = (request, response) => {
     });
 };
 
+exports.editTodo = (request, response) => {
+  if (request.body.todoId || request.body.createdAt) {
+    response.status(403).json({ message: "Not allowed to edit" });
+  }
+  const document = db.doc(`/todos/${request.params.todoId}`);
+  document
+    .get()
+    .then((doc) => {
+      if (!doc.exists) {
+        return response.status(404).json({ error: "Todo not found" });
+      }
+      if (doc.data().username !== request.user.username) {
+        return response.status(403).json({ error: "UnAuthorized process" });
+      }
+      return document.update(request.body)
+      .then(()=>{
+        response.status(202).json({message: 'Updated successfull'})
+      });      
+    })    
+    .catch((err) => {
+      console.error(err);
+      return response.status(500).json({ error: err.code });
+    });
+};
+
 exports.deleteTodo = (request, response) => {
   const document = db.doc(`/todos/${request.params.todoId}`);
   document
-      .get()
-      .then((doc) => {
-          if (!doc.exists) {
-              return response.status(404).json({ error: 'Todo not found' })
-          }
-          return document.delete();
-      })
-      .then(() => {
-          response.json({ message: 'Delete successfull' });
-      })
-      .catch((err) => {
-          console.error(err);
-          return response.status(500).json({ error: err.code });
-      });
+    .get()
+    .then((doc) => {
+      if (!doc.exists) {
+        return response.status(404).json({ error: "Todo not found" });
+      }
+      if (doc.data().username !== request.user.username) {
+        return response.status(403).json({ error: "UnAuthorized" });
+      }
+      return document.delete();
+    })
+    .then(() => {
+      response.json({ message: "Delete successfull" });
+    })
+    .catch((err) => {
+      console.error(err);
+      return response.status(500).json({ error: err.code });
+    });
 };
